AtomicTorch Studio's changes in this fork: 1. `float` -> `double` changed almost everywhere to match WPF API (so it's possible to use WPF code without any conditional compilation for NoesisGUI - but you will need to use Roslyn or any other tool to remap `System.Windows.Controls` to `Noesis` as we do in our engine for [CryoFall](CryoFall.com)). 2. `Noesis.EventArgs` -> `Noesis.NoesisEventArgs` class name changed to ensure there are no conflict between `System.EventArgs` and `Noesis.EventArgs` when we're using both in our C# files (from WPF project without any specific conditional compilation for NoesisGUI). 3. Don't use `Type.TypeHandle` property when it's possible to simply use `Type` instance. AFAIK it's used for compatibility with Mono which is not a concern for us. 4. Added our custom csproj and sln files. They don't produce NuGet package and automatically copy the `win-x64` runtime only into the solution directory `Libs/Output` (probably something which you don't want to have but you can also use it adapt for your project needs).
This fork still requires acquiring of the NoesisGUI user license and doesn't violate the original source code license.

+++ Src/Noesis/Core/Src/Proxies/BackEase.cs
@@ -42,9 +42,9 @@ public class BackEase : EasingFunctionBase {
     }
   }
 
-  public float Amplitude {
+  public double Amplitude {
     set {
-      NoesisGUI_PINVOKE.BackEase_Amplitude_set(swigCPtr, value);
+      NoesisGUI_PINVOKE.BackEase_Amplitude_set(swigCPtr, (float)value);
     } 
     get {
       float ret = NoesisGUI_PINVOKE.BackEase_Amplitude_get(swigCPtr);
